backend/requirements.txt
fastapi
uvicorn
torch
diffusers
Pillow
opencv-python
numpy
pydantic
python-dotenv
pydantic-settings 
python-multipart
gradio
transformers
accelerate